Pull in merge with minetest & irrlichtmt upstream

+++ fetch_sources.sh
@@ -29,11 +29,12 @@ getrepo zstd "https://github.com/facebook/zstd.git" e47e674cd09583ff0503f0f6defd
 
 # These repos are part of the fork
 getrepo webshims "https://github.com/paradust7/webshims.git" 91c3fe85d2cb7f85cc8e19d3f53dc8f252a69ff7
-getrepo minetest "https://github.com/paradust7/minetest.git" 08c43ff4602810269a204bcae1cdd72005b82f22
-getrepo irrlichtmt "https://github.com/paradust7/irrlicht.git" d1dae521fb75d2140e3db20a4562f907cdb77971
+getrepo minetest "https://github.com/paradust7/minetest.git" 4f32398cdf2b245bd8661f1ff8b0efc74c148cc9
+getrepo irrlichtmt "https://github.com/paradust7/irrlicht.git" ca4e12892e949b28649fc0db7e1d8d5dd71d3c66
 
 # Make irrlichtmt symlink
 pushd "$SOURCES_DIR"/minetest/lib
-rm -f irrlichtmt
-ln -s ../../irrlichtmt irrlichtmt
+if [ ! -e irrlichtmt ]; then
+  ln -s ../../irrlichtmt irrlichtmt
+fi
 popd
